mend.bbclass: Avoid code duplication for HTTP
When using the Mend API through HTTP PUT Requests, the code was being
duplicated, making the function not easily readable and inelegant.

Improve this by defining a function to access the Mend API, which
handles the HTTP request internally and returns the HTTP response.

Replace the instances of duplicated code with calls to the new function.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>