mend.wss.config: Add original defaults to excludes
A problem was encountered while mend checking in Yocto which would
result in nested duplicated libraries, and therefore vulnerabilities,
derived from a single project.
These came from the standard `.ps` directory added by the build engine
during the do_patch phase.

The cause of this issue was that the `excludes` variable, which
explicitly tells the unified_agent which files and directories to ignore
while parsing the source directory of each package, was changed from the
default value removing some parts.
Among the parts removed was "**/.*", which ignores all directories and
files starting with `.`, effectively preventing this issue.

Fix by adding back the default regexes to `excludes`, without removing
the extra ones.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>