mend.bbclass: Remove java dependency
Using the Mend Unified Agent jar forces us to rely on a custom
kas-container with Java installed, and to use HOSTTOOLS to avoid
dependency loops.

Both of these things are less than ideal.

To address this, introduce the self-contained `mend-cli` binary in the
meta layer's files, which is then directly accessed by kas-container.
This binary allows us to use the Mend Unified Agent WITHOUT having the
Java JRE and SDK installed, which means it is possible to use the
regular kas-container.
Additionally, since it is provided directly as a binary by Mend, it does
not require compilation and helps to avoid dependency loops.

Add the necessary environment variables to authenticate the mend-cli
tool, and move the mend_unified_agent command from using the jar to the
new tool.

Update the README to account for these changes.

Link: https://docs.mend.io/platform/latest/scan-with-the-mend-cli
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>